requirements.txt
Flask
requests
pandas
plotly
numpy
beautifulsoup4
openai
python-dotenv
amadeus
scikit-learn
forex-python 